# api/docs/swagger.yaml
basePath: /localeats.uz
definitions:
  dish.DishInfo:
    properties:
      allergens:
        items:
          type: string
        type: array
      available:
        type: boolean
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      dietary_info:
        items:
          type: string
        type: array
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      kitchen_name:
        type: string
      name:
        type: string
      nutrition_info:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  dish.DishShortInfo:
    properties:
      available:
        type: boolean
      category:
        type: string
      id:
        type: string
      kitchen_id:
        type: string
      kitchen_name:
        type: string
      price:
        type: number
    type: object
  dish.Dishes:
    properties:
      dishes:
        items:
          $ref: '#/definitions/dish.DishShortInfo'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dish.NutritionInfo:
    properties:
      allergens:
        items:
          type: string
        type: array
      calories:
        type: integer
      carbohydrates:
        type: integer
      dietary_info:
        items:
          type: string
        type: array
      fat:
        type: integer
      id:
        type: string
      protein:
        type: integer
    type: object
  dish.Recommendations:
    properties:
      dishes:
        items:
          $ref: '#/definitions/dish.DishShortInfo'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dish.ReqCreateDish:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      ingredients:
        items:
          type: string
        type: array
      kitchen_id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  dish.ReqUpdateDish:
    properties:
      available:
        type: boolean
      category:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
    type: object
...
  models.Error:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: LocalEats is a program to order local and homemade food with quality
    and precise delivery.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: LocalEats API Gateway
  version: "1.0"
paths:
  /dishes/{id}:
    get:
      consumes:
      - application/json
      description: gets dish by id
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.DishInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets dish by id
      tags:
      - Dishes
  /dishes/{id}/update:
    put:
      consumes:
      - application/json
      description: Update dish of kitchen
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      - description: Dish information
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/dish.ReqUpdateDish'
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.DishInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Update dish
      tags:
      - Dishes
  /dishes/{id}/updatenutritions:
    put:
      consumes:
      - application/json
      description: UpdateDishNutritionInfo
      parameters:
      - description: dish id
        in: path
        name: id
        required: true
        type: string
      - description: Dish information
        in: body
        name: kitchen
        required: true
        schema:
          $ref: '#/definitions/dish.NutritionInfo'
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.DishInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: UpdateDishNutritionInfo
      tags:
      - Dishes
  /dishes/{kitchen_id}:
    delete:
      consumes:
      - application/json
      description: deletes dish
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.Dishes'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: deletes dish
      tags:
      - Dishes
  /dishes/all/{id}:
    get:
      consumes:
      - application/json
      description: gets dishes of the kitchen
      parameters:
      - description: kitchen id
        in: path
        name: id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.Dishes'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets dishes
      tags:
      - Dishes
  /dishes/create:
    post:
      consumes:
      - application/json
      description: creates dish
      parameters:
      - description: Kitchen information
        in: body
        name: dish
        required: true
        schema:
          $ref: '#/definitions/dish.ReqCreateDish'
      produces:
      - application/json
      responses:
        "201":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.DishInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Creates dish of kitchen
      tags:
      - Dishes
  /dishes/recommendations/{user_id}:
    get:
      consumes:
      - application/json
      description: gets dishes of the kitchen
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dish informations
          schema:
            $ref: '#/definitions/dish.Recommendations'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets dishes
      tags:
      - Dishes
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"